move render result set to test
